console: Move gateway actions to new file move action creators to actions/gateways.js
