+++ pkg/webui/console/views/gateway-collaborator-edit/index.js
@@ -27,8 +27,10 @@ import Message from '../../../lib/components/message'
 import IntlHelmet from '../../../lib/components/intl-helmet'
 import toast from '../../../components/toast'
 
-import { getGatewayCollaboratorsList } from '../../store/actions/gateway'
-import { getGatewaysRightsList } from '../../store/actions/gateways'
+import {
+  getGatewayCollaboratorsList,
+  getGatewaysRightsList,
+} from '../../store/actions/gateways'
 import {
   selectSelectedGatewayId,
   selectGatewayRights,
@@ -36,6 +38,7 @@ import {
   selectGatewayRightsFetching,
   selectGatewayRightsError,
 } from '../../store/selectors/gateway'
+
 import api from '../../api'
 
 @connect(function (state, props) {
