+++ pkg/webui/console/views/gateway-collaborators-list/index.js
@@ -19,9 +19,9 @@ import { connect } from 'react-redux'
 
 import IntlHelmet from '../../../lib/components/intl-helmet'
 import CollaboratorsTable from '../../containers/collaborators-table'
-import { getGatewayCollaboratorsList } from '../../store/actions/gateway'
 import sharedMessages from '../../../lib/shared-messages'
 
+import { getGatewayCollaboratorsList } from '../../store/actions/gateways'
 import { selectSelectedGatewayId } from '../../store/selectors/gateway'
 
 const COLLABORATORS_TABLE_SIZE = 10
